refactor: 상품관리 페이지 분리
